refactor: consolidate ai-sdk-duplicate-ids tests with v5/v6 provider support
