+++ packages/memory/integration-tests/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "version": "0.1.0",
   "scripts": {
-    "test": "pnpm compile:worker && pnpm test:pg && pnpm test:upstash && pnpm test:libsql && pnpm test:streaming && pnpm test:working-memory && pnpm test:agent-memory && pnpm test:fastembed",
+    "test": "pnpm compile:worker && pnpm test:pg && pnpm test:upstash && pnpm test:libsql && pnpm test:streaming && pnpm test:working-memory && pnpm test:agent-memory && pnpm test:fastembed && pnpm test:ai-sdk-ids",
     "pretest:pg": "docker compose up -d postgres && (for i in $(seq 1 30); do docker compose exec -T postgres pg_isready -U postgres && break || (sleep 1; [ $i -eq 30 ] && exit 1); done)",
     "test:pg": "vitest run ./src/with-pg-storage.test.ts",
     "posttest:pg": "docker compose down --volumes",
@@ -26,12 +26,18 @@
     "posttest:libsql:perf": "docker compose down --volumes",
     "test:perf": "pnpm test:pg:perf && pnpm test:upstash:perf && pnpm test:libsql:perf",
     "test:fastembed": "vitest run ./src/fastembed.test.ts",
+    "test:ai-sdk-ids": "vitest run ./src/ai-sdk-duplicate-ids.test.ts",
     "compile:worker": "npx tsc ./src/worker/generic-memory-worker.ts ./src/worker/mock-embedder.ts --esModuleInterop --resolveJsonModule --module commonjs --target es2020 --outDir ./ --rootDir ./ --skipLibCheck",
     "dev": "mastra dev"
   },
   "dependencies": {
     "@ai-sdk/google": "^1.2.19",
+    "@ai-sdk/google-v5": "npm:@ai-sdk/google@^2.0.0",
+    "@ai-sdk/google-v6": "npm:@ai-sdk/google@latest",
+    "@ai-sdk/anthropic-v5": "npm:@ai-sdk/anthropic@^2.0.0",
+    "@ai-sdk/anthropic-v6": "npm:@ai-sdk/anthropic@latest",
     "@ai-sdk/openai": "^1.3.0",
+    "@ai-sdk/openai-v5": "npm:@ai-sdk/openai@^2.0.0",
     "@ai-sdk/openai-v6": "npm:@ai-sdk/openai@latest",
     "@ai-sdk/react": "^1.2.1",
     "@mastra/ai-sdk": "latest",
